+++ src/main/java/de/uni_kl/cs/discodnc/CurveBackend_MPARTC.java
@@ -1,21 +1,21 @@
-package de.uni_kl.cs.discodnc.nc;
+package de.uni_kl.cs.discodnc;
 
 import java.io.File;
 
 import de.uni_kl.cs.discodnc.CurveBackend;
 import de.uni_kl.cs.discodnc.curves.CurvePwAffine;
 import de.uni_kl.cs.discodnc.curves.LinearSegment;
-import de.uni_kl.cs.discodnc.curves.mpa_rtc_pwaffine.Curve_MPARTC_PwAffine;
-import de.uni_kl.cs.discodnc.curves.mpa_rtc_pwaffine.LinearSegment_MPARTC_PwAffine;
+import de.uni_kl.cs.discodnc.curves.mpartc.pwaffine.Curve_MPARTC_PwAffine;
+import de.uni_kl.cs.discodnc.curves.mpartc.pwaffine.LinearSegment_MPARTC_PwAffine;
 import de.uni_kl.cs.discodnc.minplus.MinPlus;
-import de.uni_kl.cs.discodnc.minplus.MinPlus_MPA_RTC;
+import de.uni_kl.cs.discodnc.minplus.mpa_rtc.MinPlus_MPARTC;
 
-public enum CurveBackend_MPA_RTC implements CurveBackend {
-	MPA_RTC;
+public enum CurveBackend_MPARTC implements CurveBackend {
+	MPARTC;
 
 	@Override
 	public MinPlus getMinPlus() {
-		return MinPlus_MPA_RTC.MINPLUS_MPA_RTC;
+		return MinPlus_MPARTC.MINPLUS_MPARTC;
 	}
 
 	@Override
